chg:dev:support loading only the declared openstack modules using an environment variable
By default, osclients import all the openstack modules. This implies that all the python-*client must be added to requirements, even when they are not used.

To avoid this, the new environment variable OSCLIENTS_MODULES may be defined.

If defined, it is a list of modules, separated by comma. Possible values are nova, glance, neutron, cinder, swift. Keystone is not included in the list because it is always required.